Ensure turborepo tests run under swc
