+++ apps/slack-bot/jest.config.js
@@ -1,18 +1,15 @@
-const { readFileSync } = require('fs');
-
-const swcJestConfig = JSON.parse(
-  readFileSync(`${__dirname}/.spec.swcrc`, 'utf-8'),
-);
-
-swcJestConfig.swcrc = false;
+const swcJestConfig = {
+  ...require('../../jest.swc.config.cjs'),
+  swcrc: false,
+};
 
 module.exports = {
   displayName: 'slack-bot',
-  preset: '../../jest.preset.js',
   testEnvironment: 'node',
   transform: {
     '^.+\\.[tj]s$': ['@swc/jest', swcJestConfig],
   },
+  transformIgnorePatterns: ['node_modules/(?!(graphql-request)/)'],
   moduleFileExtensions: ['ts', 'js', 'html'],
   coverageDirectory: 'test-output/jest/coverage',
   collectCoverageFrom: [
