+++ apps/slack-bot/tsconfig.json
@@ -16,7 +16,7 @@
   "exclude": [
     "out-tsc",
     "dist",
-    "jest.config.ts",
+    "jest.config.js",
     "src/**/*.spec.ts",
     "src/**/*.test.ts",
     "eslint.config.js",
