--- libs/redis-client/jest.config.ts
@@ -1,23 +1,0 @@
-/* eslint-disable */
-import { readFileSync } from 'fs';
-
-// Reading the SWC compilation config for the spec files
-const swcJestConfig = JSON.parse(
-  readFileSync(`${__dirname}/.spec.swcrc`, 'utf-8'),
-);
-
-// Disable .swcrc look-up by SWC core because we're passing in swcJestConfig ourselves
-swcJestConfig.swcrc = false;
-
-export default {
-  displayName: '@mud/redis-client',
-  preset: '../../jest.preset.js',
-  testEnvironment: 'node',
-  transform: {
-    '^.+\\.[tj]s$': ['@swc/jest', swcJestConfig],
-  },
-  moduleFileExtensions: ['ts', 'js', 'html'],
-  transformIgnorePatterns: ['/node_modules/(?!(graphql-request)/)'],
-  coverageDirectory: 'test-output/jest/coverage',
-  collectCoverageFrom: ['src/**/*.{ts,js}'],
-};
