+++ package.json
@@ -4,19 +4,26 @@
   "license": "MIT",
   "private": true,
   "scripts": {
-    "build": "turbo run build",
-    "db:migrate:deploy": "turbo run db:migrate:deploy",
-    "db:migrate:dev": "turbo run db:migrate:dev",
-    "db:push": "turbo run db:push",
-    "test": "turbo run test",
-    "db:seed": "turbo run db:seed",
-    "serve": "turbo run serve",
+    "turbo": "node ./node_modules/turbo/bin/turbo",
+    "build": "yarn turbo run build",
+    "db:migrate:deploy": "yarn turbo run db:migrate:deploy",
+    "db:migrate:dev": "yarn turbo run db:migrate:dev",
+    "db:push": "yarn turbo run db:push",
+    "test": "yarn turbo run test",
+    "db:seed": "yarn turbo run db:seed",
+    "serve": "yarn turbo run serve",
     "format": "prettier --write \"**/*.{ts,tsx,md}\"",
-    "generate": "turbo run generate",
-    "lint": "turbo run lint"
+    "generate": "yarn turbo run generate",
+    "lint": "yarn turbo run lint"
   },
   "devDependencies": {
+    "@nx/jest": "^20.4.3",
+    "@swc/core": "^1.10.12",
+    "@swc/jest": "^0.2.36",
+    "@types/jest": "^29.5.13",
+    "jest": "^29.7.0",
     "prettier": "^3.6.2",
+    "ts-jest": "^29.2.5",
     "turbo": "^2.5.8"
   },
   "dependencies": {},
